fixup! rm config, _app -> lib/*
