--- config/env_dev.json
@@ -1,4 +1,0 @@
-{
-  "AUTH0_KEY": "DWrTFyyzp3QZq8S3PZ5dfJd3T8xNf5kY",
-  "AUTH0_URL": "publicbits.auth0.com"
-}
