--- config/env_prod.json
@@ -1,3 +1,0 @@
-{
-  "key": "value"
-}
